BackEnd/voice_model/requirements.txt
# Voice Emotion Detection API Requirements
# Core dependencies for the FastAPI voice emotion detection service

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and Audio Processing
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1
ffmpeg-python==0.2.0

# Data handling and utilities
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.1.3

# HTTP and CORS
httpx==0.25.2
pydantic==2.5.0

# Audio format support
soundfile==0.12.1
resampy==0.4.2

# Development and testing
pytest==7.4.3
requests==2.31.0

# Spotify API integration
spotipy==2.23.0
python-dotenv==1.0.0

# GPU support (CUDA 12.1 compatible)
torch==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.20.1+cu121 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121